fix: Move coprocessor.rs to asm.rs as requested
